+++ common/GM_APP/Src/app_pid_init_cmd.c
@@ -67,13 +67,13 @@ __weak void app_rx_init_sub_pid_driver_data1_rqt(uint8_t num, prtc_header_t *pPh
 	
 }
 
-void app_tx_init_sub_pid_driver_data1_ctl(uint8_t num, uint8_t priority, uint8_t souce_id, uint8_t target_id, uint8_t sub_id, uint8_t direction, uint16_t angle, uint16_t init_position, uint8_t reducer_ratio)
+void app_tx_init_sub_pid_driver_data1_ctl(uint8_t num, uint8_t priority, uint8_t souce_id, uint8_t target_id, uint8_t sub_id, uint8_t direction, uint16_t angle, uint16_t init_position, uint16_t reducer_ratio)
 {
 	hal_can_protocol_tx(num, net_pha(CMD_CONTROL, sizeof(prtc_data_ctl_init_driver_data1_t), priority, souce_id, target_id, PID_INIT, INIT_SUB_PID_DRIVER_DATA1, sub_id), \
 			net_pda_init_sub_pid_driver_data1_ctl(direction, angle, init_position, reducer_ratio));
 }
 
-void app_tx_init_sub_pid_driver_data1_rsp(uint8_t num, uint8_t priority, uint8_t souce_id, uint8_t target_id, uint8_t sub_id, uint8_t direction, uint16_t angle, uint16_t init_position, uint8_t reducer_ratio)
+void app_tx_init_sub_pid_driver_data1_rsp(uint8_t num, uint8_t priority, uint8_t souce_id, uint8_t target_id, uint8_t sub_id, uint8_t direction, uint16_t angle, uint16_t init_position, uint16_t reducer_ratio)
 {
 	hal_can_protocol_tx(num, net_pha(CMD_RESPONSE, sizeof(prtc_data_rsp_init_driver_data1_t), priority, souce_id, target_id, PID_INIT, INIT_SUB_PID_DRIVER_DATA1, sub_id), \
 			net_pda_init_sub_pid_driver_data1_rsp(direction, angle, init_position, reducer_ratio));
